requirements.txt
django-memoize==1.2.0
django-redis==4.2.0
Django==1.9.9
Jinja2==2.8
psycopg2==2.6.1
schematics==1.1.1
ujson==1.35
